Added EditorCameraHelper
